added visuals

+++ apacheTVMSpace/matMul_XGBTuner.py
@@ -36,7 +36,6 @@ def matmul_v1(N, L, M, dtype):
     cfg.define_knob("tile_y", [2,4,8,16,32,64,128,256,512,1024])
     cfg.define_knob("tile_x", [2,4,8,16,32,64,128,256,512,1024])
 
-
     # 4. schedule according to config
     yo, yi = s[C].split(y, cfg["tile_y"].val)
     xo, xi = s[C].split(x, cfg["tile_x"].val)
@@ -71,7 +70,7 @@ def main():
 
     with autotvm.apply_history_best("results/tvm_XGBTuner.json"):
         with tvm.target.Target("llvm"):
-            s, arg_bufs = matmul_v1(N, L, M,"float64") #float64 
+            s, arg_bufs = matmul_v1(N, L, M,"float32") #float64 
             func = tvm.build(s, arg_bufs)
 
 if __name__ == '__main__':
